src/main/java/org/avlasov/leetcodeproblems/easy/CountPrimes.java
package org.avlasov.leetcodeproblems.easy;

/**
 * Created By artemvlasov on 16/07/2018
 *
 * https://leetcode.com/problems/count-primes/description/
 *
 * Count the number of prime numbers less than a non-negative number, n.
 *
 * Example:
 *
 * Input: 10
 * Output: 4
 * Explanation: There are 4 prime numbers less than 10, they are 2, 3, 5, 7.
 *
 **/
public class CountPrimes {

    public int countPrimes(int n) {
        if (n <= 1) return 0;
        boolean[] notPrime = new boolean[n];
        int count = 0;
        for (int i = 2; i < n; i++) {
            if (!notPrime[i]) {
                count++;
                for (int j = 2; i*j < n; j++) {
                    notPrime[i*j] = true;
                }
            }
        }
        return count;
    }

}
